[FEAT] Use global flag for mass messaging
The global flag on the notifications table was not being used.  I'm
changing this to be the signal that this notification is one of a
mass set, which can then be used in conjunction with the
mailboxer_email method on the messageable model to determine if a
message should be emailed, or not.